wip:헤더 컴포넌트에서 사용하는 쿼리파람 로직을 커스텀 훅으로 분리 완료
